Code/requirements.txt
requests
Flask
mysql-connector-python
confluent-kafka
Flask_SocketIO
python-socketio
simple-websocket
asyncio
websockets
flask-sock
